#11867 Better warehouse support and inventory control. Now a single product can exist in multiple warehouses.
Note 1: Please note that in this case inventory is adjusted when a shipment is created or deleted by a store owner. In other cases inventory is adjusted when order is placed/cancelled/deleted.

Note 2: a lot of changes and enhacements are coming (e.g. reserved qty, work items #12278 and #12279, etc)